Create separate file for create_genre
